Extract binary manipulations from android into shared module.
