+++ zeapp/android/src/main/java/de/berlindroid/zeapp/zebits/ZeStaticPattern.kt
@@ -1,5 +1,9 @@
 package de.berlindroid.zeapp.zebits
 
+import de.berlindroid.zekompanion.grayscale
+import de.berlindroid.zekompanion.green
+import de.berlindroid.zekompanion.map
+import de.berlindroid.zekompanion.rgb
 import java.nio.IntBuffer
 import kotlin.math.abs
