+++ zeapp/android/src/main/java/de/berlindroid/zeapp/zeservices/ZeBadgeManager.kt
@@ -3,13 +3,13 @@ package de.berlindroid.zeapp.zeservices
 import android.content.Context
 import android.graphics.Bitmap
 import dagger.hilt.android.qualifiers.ApplicationContext
-import de.berlindroid.zeapp.zebits.base64
-import de.berlindroid.zeapp.zebits.toBinary
-import de.berlindroid.zeapp.zebits.zipit
 import de.berlindroid.zeapp.zeui.pixelBuffer
 import de.berlindroid.zekompanion.BadgePayload
 import de.berlindroid.zekompanion.Environment
+import de.berlindroid.zekompanion.base64
 import de.berlindroid.zekompanion.buildBadgeManager
+import de.berlindroid.zekompanion.toBinary
+import de.berlindroid.zekompanion.zipit
 import javax.inject.Inject
 
 class ZeBadgeManager @Inject constructor(
