+++ zeapp/android/src/main/java/de/berlindroid/zeapp/zeui/ZeImageDrawEditorDialog.kt
@@ -39,7 +39,7 @@ import de.berlindroid.zeapp.PAGE_WIDTH
 import de.berlindroid.zeapp.R
 import de.berlindroid.zeapp.zebits.isBinary
 import de.berlindroid.zeapp.zebits.scaleIfNeeded
-import de.berlindroid.zeapp.zebits.threshold
+import de.berlindroid.zekompanion.threshold
 import de.berlindroid.zeapp.zemodels.ZeConfiguration
 import kotlinx.coroutines.launch
